added support for 65536 types of blocks

+++ worldloader.cpp
@@ -68,7 +68,7 @@ static bool loadAllRegions();
 static bool loadRegion(const char* file, const bool mustExist, int &loadedChunks);
 static bool loadTerrainRegion(const char *fromPath, int &loadedChunks);
 static bool scanWorldDirectoryRegion(const char *fromPath);
-static inline void assignBlock(const uint8_t &source, uint8_t* &dest, int &x, int &y, int &z, uint8_t* &justData);
+static inline void assignBlock(const uint16_t &source, uint16_t* &dest, int &x, int &y, int &z, uint8_t* &justData);
 static inline void lightCave(const int x, const int y, const int z);
 
 int getWorldFormat(const char *worldPath)
@@ -443,7 +443,7 @@ static bool loadChunk(const char *streamOrFile, const size_t streamLen)
 					*bp-- = AIR;
 				}
 			}
-			uint8_t *targetBlock;
+			uint16_t *targetBlock;
 			if (g_Orientation == East) {
 				targetBlock = &BLOCKEAST(x + offsetx, 0, z + offsetz);
 			} else if (g_Orientation == North) {
@@ -557,7 +557,7 @@ static bool loadAnvilChunk(NBT_Tag * const level, const int32_t chunkX, const in
 		// Copy data
 		for (int x = 0; x < CHUNKSIZE_X; ++x) {
 			for (int z = 0; z < CHUNKSIZE_Z; ++z) {
-				uint8_t *targetBlock, *lightByte;
+				uint16_t *targetBlock, *lightByte;
 				if (g_Orientation == East) {
 					targetBlock = &BLOCKEAST(x + offsetx, yoffset, z + offsetz);
 					if (g_Skylight || g_Nightmode) lightByte = &SETLIGHTEAST(x + offsetx, yoffset, z + offsetz);
@@ -770,12 +770,12 @@ static void allocateTerrain()
 	memset(g_HeightMap, 0, g_MapsizeX * g_MapsizeZ * sizeof(uint16_t));
 	const size_t terrainsize = g_MapsizeZ * g_MapsizeX * g_MapsizeY;
 	printf("Terrain takes up %.2fMiB", float(terrainsize / float(1024 * 1024)));
-	g_Terrain = new uint8_t[terrainsize];
+	g_Terrain = new uint16_t[terrainsize];
 	memset(g_Terrain, 0, terrainsize); // Preset: Air
 	if (g_Nightmode || g_Underground || g_BlendUnderground || g_Skylight) {
 		lightsize = g_MapsizeZ * g_MapsizeX * ((g_MapsizeY + (g_MapminY % 2 == 0 ? 1 : 2)) / 2);
 		printf(", lightmap %.2fMiB", float(lightsize / float(1024 * 1024)));
-		g_Light = new uint8_t[lightsize];
+		g_Light = new uint16_t[lightsize];
 		// Preset: all bright / dark depending on night or day
 		if (g_Nightmode) {
 			memset(g_Light, 0x11, lightsize);
@@ -1028,7 +1028,7 @@ static void loadBiomeChunk(const char* path, const int chunkX, const int chunkZ)
 	delete[] file;
 }
 
-static inline void assignBlock(const uint8_t &block, uint8_t* &targetBlock, int &x, int &y, int &z, uint8_t* &justData)
+static inline void assignBlock(const uint16_t &block, uint16_t* &targetBlock, int &x, int &y, int &z, uint8_t* &justData)
 {
 	if (block == WOOL || block == LOG || block == LEAVES || block == STEP || block == DOUBLESTEP || block == WOOD || block == WOODEN_STEP || block == WOODEN_DOUBLE_STEP 
 		|| block == 95 || block == 160 || block == 159 || block == 171 || block == 38 || block == 175 || block == SAND || block == 153
@@ -1309,7 +1309,7 @@ void uncoverNether()
 		for (size_t z = CHUNKSIZE_Z; z < g_MapsizeZ - CHUNKSIZE_Z; ++z) {
 			// Remove blocks on top, otherwise there is not much to see here
 			int massive = 0;
-			uint8_t *bp = g_Terrain + ((z + (x * g_MapsizeZ) + 1) * g_MapsizeY) - 1;
+			uint16_t *bp = g_Terrain + ((z + (x * g_MapsizeZ) + 1) * g_MapsizeY) - 1;
 			int i;
 			for (i = 0; i < to; ++i) { // Go down 74 blocks from the ceiling to see if there is anything except solid
 				if (massive && (*bp == AIR || *bp == LAVA || *bp == STAT_LAVA)) {
